+++ jams/minimal/ex_minimal.h
@@ -17,9 +17,9 @@ extern "C"
 
 namespace mud
 {
-	template <> struct TypedBuffer<Bullet> { static size_t index() { return 20; } };
-	template <> struct TypedBuffer<Human>  { static size_t index() { return 21; } };
-	template <> struct TypedBuffer<Crate>  { static size_t index() { return 22; } };
+	template <> struct TypedBuffer<Bullet> { static uint32_t index() { return 20; } };
+	template <> struct TypedBuffer<Human>  { static uint32_t index() { return 21; } };
+	template <> struct TypedBuffer<Crate>  { static uint32_t index() { return 22; } };
 }
 
 namespace mud
